moving write_dir to conftest
